Objective

WaterModels.objective_des — Method

objective_des(wm::AbstractWaterModel)

Sets the objective function for network design (des) problem specifications. By default, the cost of selecting discrete network resistances is minimized.

WaterModels.objective_wf — Method

objective_wf(wm::AbstractWaterModel)

Sets the objective function for Water Flow (WF) problem specifications. By default, only feasibility must be satisfied.

Objective Functions

WaterModels.objective_owf — Function

Defines the objective for the owf problem is NLP formulations.